pyproject.toml
[project]
name = "before-you-sign"
version = "0.1.0"
description = "Assistant for reading EULA, ToS, etc."
authors = [
    {name = "Peter Minin", email = "peterminin@gmail.com"},
]
dependencies = [
    "gradio~=5.6.0",
    "pypandoc-binary~=1.14",
    "google-generativeai~=0.8.3",
    "pyyaml~=6.0.2",
]
requires-python = "~=3.12"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = true
